app/pyproject.toml
[tool.poetry]
name = "lala"
version = "0.1.0"
description = ""
authors = ["Zalim"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.8.0"
python-dotenv = "^1.0.1"
redis = "^5.0.7"
uvicorn = "^0.30.1"
python-decouple = "^3.8"
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
aioredis = "^2.0.1"
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.2"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "./app/"